feat: month select for analytics page
