+++ src/routes/_app.dashboard.analytics.tsx
@@ -1,9 +1,33 @@
-import { createFileRoute } from '@tanstack/react-router'
+import { MonthSelect } from "@/components/custom/month-select";
+import { PageHeader } from "@/components/custom/page-header";
+import { createFileRoute } from "@tanstack/react-router";
+import { zodValidator } from "@tanstack/zod-adapter";
+import z from "zod";
 
-export const Route = createFileRoute('/_app/dashboard/analytics')({
+const analyticsSearchSchema = z.object({
+  month: z.number().optional(),
+  year: z.number().optional(),
+});
+
+export const Route = createFileRoute("/_app/dashboard/analytics")({
+  loaderDeps: ({ search: { month, year } }) => ({ month, year }),
+  loader: async ({ context, deps }) => {
+    // TODO: Load analytics data
+    console.log(context.user.name, deps.month, deps.year);
+  },
+  validateSearch: zodValidator(analyticsSearchSchema),
   component: RouteComponent,
-})
+});
 
 function RouteComponent() {
-  return <div>Hello "/_app/dashboard/analytics"!</div>
+  return (
+    <div className="space-y-6">
+      <PageHeader title="Analytics">
+        <MonthSelect
+          from="/_app/dashboard/analytics"
+          to="/dashboard/analytics"
+        />
+      </PageHeader>
+    </div>
+  );
 }
